Add standalone tests for 10plus6
